Archaic-Wrapper rule and visitTask capabilities
Added visitTask to the AST visitor to be invoked for every task definition.
Created the archaic-wrapper rule, supports defining a threshold for the major version
and the minor version separately - defaults to 0 for major and 2 for minor.
A task for later commits is to design a way to configure a rule from within
the linted build.