Added support for custom wrapper task names
Previous code supported only tasks named "wrapper" due to a bug in
ArchaicWrapperRule.visitExtensionProperty and the behavior of the AST.
In cases of which the task is configured separately from the declaration, the evaluation order
is task -> configuration, but when the task is declared and configured inline the evaluation order
is configuration -> task, thus we cannot assume we will have the right extension name (the task name)
present at the wrapperTaskName member when visiting extension property. So instead, keep all the potential relevant
configurations until visitClassComplete.
This will allow us to also catch overriding configurations of the same wrapper task as violations.