Regression in v6.0.0 refactoring
I'm not 100% sure if this fix keeps the same semantics as it used to be. Maybe there should be the same mapping to warn, error etc (level Enum options) . But at least in our case the "duplicate-dependency-class" rule crashed lintGradle task when there very validation errors. I don't have more time to think about it right now, but I'm certain that there could be cleaner fix for this bug.