Documentation about Kotlin build scripts
As Gradle Lint Plugin doesn't support Kolin build scripts you should convert those build scripts to Groovy.
When this is not possible you can either not use this plugin or use the "workaround" to make the plugin available on the remaining build scripts (Groovy only).
I myself searched for such a workaround and got no answer or didn't find anything. So to help people like me this was added to the main README.md!